BSERVDEV-13374: Throw if Bitbucket Server contains no release tags
The Bitbucket Server repository should always contain tags on the Maven commit for each release. If the STASH/stash linked repository in Bamboo is configured to only do a shallow clone, then these tags will not show up. Currently, the script will continue and misleadinging report that there is no new version to update, even though this may be not true. This change will cause the script to fail early if the tags list is empty (which shouldn't ever happen if the linked repository is configured correctly).